Move build tool related classes to new projects.buildtools package
