+++ cli/ballerina-cli/src/main/java/module-info.java
@@ -1,11 +1,9 @@
 module io.ballerina.cli {
     uses io.ballerina.cli.BLauncherCmd;
-    uses io.ballerina.cli.tool.CodeGeneratorTool;
     exports io.ballerina.cli;
     exports io.ballerina.cli.launcher;
     exports io.ballerina.cli.utils;
     exports io.ballerina.cli.cmd;
-    exports io.ballerina.cli.tool;
 
     requires io.ballerina.runtime;
     requires io.ballerina.lang;
